pyproject.toml
[project]
name = "agent-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.7",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "mcp>=1.27.0",
    "openai>=2.31.0",
    "openai-agents>=0.13.6",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "rich>=15.0.0",
    "tenacity>=9.1.4",
    "wikipedia-api>=0.14.1",
]
